Cambios parciales para turnosAsignados de paciente
